# docs/swagger.yaml
definitions:
  internal.CreateUserRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword123
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal.CreateUserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal.ErrorResponse:
    properties:
      details:
        example: Email is required
        type: string
      error:
        example: Validation failed
        type: string
    type: object
  internal.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  internal.LogoutRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  internal.MessageResponse:
    properties:
      message:
        example: Successfully logged out
        type: string
    type: object
  internal.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  internal.RefreshResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
info:
  contact: {}
paths:
  /identity/logout:
    post:
      consumes:
      - application/json
      description: Выполняет выход пользователя и блокирует сессию
      parameters:
      - description: Refresh token для выхода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Выход из системы
      tags:
      - identity
  /identity/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены с использованием валидного refresh
        токена
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Обновление токенов
      tags:
      - identity
  /identity/sign-in:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя и возвращает токены доступа
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - identity
  /identity/sign-up:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - identity
swagger: "2.0"